# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/logo:
    post:
      tags: [Storage]
      operationId: uploadLogo
      summary: Upload a clinic logo (server-proxied)
      description: |
        Accepts raw image bytes (JPEG, PNG, GIF, WebP, SVG) up to 2 MB.
        The server validates MIME type and size before writing to GCS, ensuring
        constraints are enforced at the actual ingestion boundary.
        Requires Authorization: Bearer <admin-jwt>.
      requestBody:
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Logo stored; returns the object path to save in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoUploadResponse'
        '400':
          description: Invalid MIME type, empty body, or file too large.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid admin JWT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: File exceeds 2 MB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to upload logo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves clinic logo objects from the private object dir.
        Only the `logos/` prefix is publicly accessible; all other paths return 403.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the logos namespace (e.g. `logos/some-uuid`).
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    LogoUploadResponse:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/logos/uuid`). Store this in the database.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string