# lib/api-zod/src/generated/types/logoUploadResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Clinic OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogoUploadResponse {
  /** Normalized object path (e.g. `/objects/logos/uuid`). Store this in the database. */
  objectPath: string;
}